ai/requirements.txt
absl-py==2.1.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
html5lib==1.1
idna==3.10
jax==0.4.30
jaxlib==0.4.30
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
ml_dtypes==0.5.1
multitasking==0.0.11
newsapi-python==0.2.7
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
praw==7.8.1
prawcore==2.4.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tweepy==4.15.0
typing_extensions==4.5.0
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.14.1
yfinance==0.2.54
